Removed 2xSaI.cpp and hq2x.cpp from scale32 MSVC project. Further optimized image blitting routines in scale16/32 and standard16/32.
git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@1316 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2

+++ sphere/source/video/scale16/scale16.cpp
@@ -1277,7 +1277,7 @@ void NormalBlit(IMAGE image, int x, int y)
 {
     calculate_clipping_metrics(image->width, image->height);
 
-    byte a;
+    int a;
     word result;
 
     word* dst   = (word*)ScreenBuffer + (y + image_offset_y) * ScreenWidth  + image_offset_x + x;
